+++ LEGACY/SirThaddeus.DesktopRuntime/Controls/PersonalitySettingsControl.xaml.cs
@@ -1,9 +1,0 @@
-namespace SirThaddeus.DesktopRuntime.Controls;
-
-public partial class PersonalitySettingsControl : System.Windows.Controls.UserControl
-{
-    public PersonalitySettingsControl()
-    {
-        InitializeComponent();
-    }
-}