+++ LEGACY/SirThaddeus.DesktopRuntime/Views/HeroCardControl.xaml.cs
@@ -1,9 +1,0 @@
-namespace SirThaddeus.DesktopRuntime.Views;
-
-public partial class HeroCardControl : System.Windows.Controls.UserControl
-{
-    public HeroCardControl()
-    {
-        InitializeComponent();
-    }
-}