+++ LEGACY/SirThaddeus.DesktopRuntime/Views/MapPaneControl.xaml.cs
@@ -1,9 +1,0 @@
-namespace SirThaddeus.DesktopRuntime.Views;
-
-public partial class MapPaneControl : System.Windows.Controls.UserControl
-{
-    public MapPaneControl()
-    {
-        InitializeComponent();
-    }
-}